+++ lib/fog/azurerm/models/network/express_route_circuit_peering.rb
@@ -25,34 +25,33 @@ module Fog
         attribute :routing_registry_name
 
         def self.parse(circuit_peering)
-          circuit_peering_properties = circuit_peering['properties']
           hash = {}
-          hash['id'] = circuit_peering['id']
-          hash['name'] = circuit_peering['name']
-          hash['resource_group'] = get_resource_group_from_id(circuit_peering['id'])
-          hash['circuit_name'] = circuit_peering['id'].split('/')[8]
-          hash['provisioning_state'] = circuit_peering_properties['provisioningState']
-          hash['peering_type'] = circuit_peering_properties['peeringType']
-          hash['peer_asn'] = circuit_peering_properties['peerASN']
-          hash['azure_asn'] = circuit_peering_properties['azureASN']
-          hash['primary_azure_port'] = circuit_peering_properties['primaryAzurePort']
-          hash['secondary_azure_port'] = circuit_peering_properties['secondaryAzurePort']
-          hash['state'] = circuit_peering_properties['state']
-          hash['primary_peer_address_prefix'] = circuit_peering_properties['primaryPeerAddressPrefix']
-          hash['secondary_peer_address_prefix'] = circuit_peering_properties['secondaryPeerAddressPrefix']
-          hash['vlan_id'] = circuit_peering_properties['vlanId']
+          hash['id'] = circuit_peering.id
+          hash['name'] = circuit_peering.name
+          hash['resource_group'] = get_resource_group_from_id(circuit_peering.id)
+          hash['circuit_name'] = circuit_peering.id.split('/')[8]
+          hash['provisioning_state'] = circuit_peering.provisioning_state
+          hash['peering_type'] = circuit_peering.peering_type
+          hash['peer_asn'] = circuit_peering.peer_asn
+          hash['azure_asn'] = circuit_peering.azure_asn
+          hash['primary_azure_port'] = circuit_peering.primary_azure_port
+          hash['secondary_azure_port'] = circuit_peering.secondary_azure_port
+          hash['state'] = circuit_peering.state
+          hash['primary_peer_address_prefix'] = circuit_peering.primary_peer_address_prefix
+          hash['secondary_peer_address_prefix'] = circuit_peering.secondary_peer_address_prefix
+          hash['vlan_id'] = circuit_peering.vlan_id
 
-          microsoft_peering_config = circuit_peering_properties['microsoftPeeringConfig']
+          microsoft_peering_config = circuit_peering.microsoft_peering_config
           unless microsoft_peering_config.nil?
-            public_prefixes = microsoft_peering_config['advertisedpublicprefixes']
+            public_prefixes = microsoft_peering_config.advertised_public_prefixes
             hash['advertised_public_prefixes'] = []
             public_prefixes.each do |public_prefix|
               hash['advertised_public_prefixes'] << public_prefix
             end unless public_prefixes.nil?
 
-            hash['advertised_public_prefix_state'] = microsoft_peering_config['advertisedPublicPrefixState']
-            hash['customer_asn'] = microsoft_peering_config['customerAsn']
-            hash['routing_registry_name'] = microsoft_peering_config['routingRegistryName']
+            hash['advertised_public_prefix_state'] = microsoft_peering_config.advertised_public_prefixes_state
+            hash['customer_asn'] = microsoft_peering_config.customer_asn
+            hash['routing_registry_name'] = microsoft_peering_config.routing_registry_name
           end
           hash
         end
@@ -62,7 +61,7 @@ module Fog
           requires :advertised_public_prefixes if peering_type.casecmp(MICROSOFT_PEERING) == 0
           circuit_peering_parameters = express_route_circuit_peering_params
           circuit_peering = service.create_or_update_express_route_circuit_peering(circuit_peering_parameters)
-          merge_attributes(Fog::Network::AzureRM::ExpressRouteCircuit.parse(circuit_peering))
+          merge_attributes(Fog::Network::AzureRM::ExpressRouteCircuitPeering.parse(circuit_peering))
         end
 
         def destroy