+++ test/api_stub/models/storage/data_disk.rb
@@ -17,11 +17,11 @@ module ApiStub
 
         def self.expected_create_data_disk_response
           {
-            'lun' => 0,
-            'name' => 'disk1',
-            'vhd_uri' => 'https://confizrg7443.blob.core.windows.net/vhds/disk1.vhd',
-            'create_option' => 'empty',
-            'disk_size_gb' => '10'
+              'name' => 'disk1',
+              'disk_size_gb' => '10',
+              'lun' => 0,
+              'vhd_uri' => 'https://confizrg7443.blob.core.windows.net/vhds/disk1.vhd',
+              'create_option' => 'empty'
           }
         end
       end